test(api): rearrange test suite into single binary
