+++ tests/api/main.rs
@@ -1,0 +1,6 @@
+mod basetest;
+mod collect_xpubs_test;
+mod create_user_test;
+mod generate_address_test;
+mod masterkeys_test;
+mod ping_test;
