+++ tests/api/masterkeys_test.rs
@@ -1,12 +1,10 @@
-pub mod basetest;
-
-use basetest::base;
+use crate::basetest::{spawn_app, MasterKeysResponse};
 use std::collections::HashMap;
 
 #[tokio::test]
 async fn masterkeys_returns_200_for_valid_networks() {
     // 1. Arrange
-    let test_app = base::spawn_app().await;
+    let test_app = spawn_app().await;
     let client = reqwest::Client::new();
     let networks = vec!["bitcoin", "signet", "regtest", "testnet"];
     let url = format!("{}/masterkeys", &test_app.address);
@@ -31,7 +29,7 @@ async fn masterkeys_returns_200_for_valid_networks() {
 #[tokio::test]
 async fn masterkeys_returns_400_for_invalid_networks() {
     // 1. Arrange
-    let test_app = base::spawn_app().await;
+    let test_app = spawn_app().await;
     let client = reqwest::Client::new();
     let invalid_networks = vec!["mainnet", "segnet", "ethernet", "internet"];
     let url = format!("{}/masterkeys", &test_app.address);
@@ -48,7 +46,7 @@ async fn masterkeys_returns_400_for_invalid_networks() {
             .await
             .expect("Failed to execute request");
 
-        let resp_body = resp.json::<base::MasterKeysResponse>().await.unwrap();
+        let resp_body = resp.json::<MasterKeysResponse>().await.unwrap();
         // 3. Assert
         assert_eq!(400, resp_body.status);
         assert_eq!(None, resp_body.data);
